requirements.txt
async-generator==1.10
attrs==22.2.0
cffi==1.15.1
cfscrape==2.1.1
charset-normalizer==3.0.1
cryptography==39.0.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
requests==2.28.2
selenium==4.6.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.25.9
urllib3-secure-extra==0.1.0
wincertstore==0.2
wsproto==1.2.0
web3==5.11.1
python-binance==1.0.16
